functions/events/storage-firestore/delete/requirements.txt
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0